Initial changes to code following short review
Added and edited .gitignore to avoid pushin unnecessary files.
Cleaned up `binary_ask.py` a bit with some comments here and there.
The other files will likely need to be reworked and/or streamlined.
For example, the binary versions of ASK and FSK should be a special
case of MASK and MFSK.
The Filters file could be removed as it is a very thin wrapper around
`scipy.signal`.